api/requirements.txt
tensorflow==2.16.2
fastapi
uvicorn
python-miltipart
pillow
tensorflow-serving-api=2.16.2
matplotlib
numpy
